Removes duplicate of cs_filter
