+++ varats/plots/blame_interaction_degree.py
@@ -18,7 +18,7 @@ from varats.data.revisions import get_processed_revisions_files
 from varats.data.reports.blame_report import (BlameReport,
                                               generate_degree_tuples)
 from varats.plots.plot_utils import check_required_args
-from varats.paper.case_study import CaseStudy
+from varats.paper.case_study import CaseStudy, get_case_study_file_name_filter
 
 
 def _build_interaction_table(report_files: tp.List[Path],
@@ -58,21 +58,9 @@ def _gen_blame_interaction_data(**kwargs: tp.Any) -> pd.DataFrame:
     result_dir = Path(kwargs["result_folder"])
     project_name = kwargs["project"]
 
-    def cs_filter(file_name: str) -> bool:
-        """
-        Filter files that are not in the case study.
-
-        Returns True if a case_study is set and the commit_hash of the file
-        is not part of this case_study, otherwise, False.
-        """
-        if case_study is None:
-            return False
-
-        commit_hash = BlameReport.get_commit_hash_from_result_file(file_name)
-        return not case_study.has_revision(commit_hash)
-
-    report_files = get_processed_revisions_files(project_name, result_dir,
-                                                 BlameReport, cs_filter)
+    report_files = get_processed_revisions_files(
+        project_name, result_dir, BlameReport,
+        get_case_study_file_name_filter(case_study))
 
     data_frame = _build_interaction_table(report_files,
                                           str(project_name))
